refactor: rename PascalCase files to kebab-case in examples/hono (#129)
Rename Java-style PascalCase source files to TS-convention kebab-case:
- DatabaseHealthIndicator.ts → database-health-indicator.ts
- DatabaseModule.ts → database-module.ts
- TodoController.ts → todo-controller.ts
- TodoRepository.ts → todo-repository.ts
- TodoService.ts → todo-service.ts

Updates hand-written imports. Generated files update automatically on build.

Closes #129
