+++ examples/hono/src/database-health-indicator.ts
